Added; Shamir Shares To Registration
Basically this feature just allows for the shamir to geenrate when a user firsat regsiters, prompts them to staore the 5 secrets, as it does generate 5 secrets then it is used in the recovery of the account (making this so that you will be able to recover password to the account). but not yet.

This then allows for the shamir to be generated, encoded and decoded using cecksum validations. If successful then success, this should then reveal the password to you, is what i want to implement next.