Added: Logout Functionality
This was done using the activity_log collection table from the database, and as well, we created db for login and registration as well and for biometrics. In order to do so for biometrircs, an api route was needed to be made for biometric login, we had one for registration but it wasnt going to work for login, and the biometric login, was used in the client side, so basically we needed to make a route. Changed the dummy activites using database collected activities. And for logout and logs route we needed to use authorization headers for the user token to decode and decode it